Add new Async sign method to TransactionEncoder
Add async methods to allow for out of process signing to use efficient
non-blocking constructions.

We use a double-barrelled continuation passing style to allow for both
sync and async code to share the majority of code with minimal overhead
whhile still leaving the door open to the user overriding the
async methods with their framework of choice (Futures etc).