Web3a: Introduce compatibility submodule to deal with Android integration
This commit adds JDK jars used by Android during compilation:
android.jar and core-lambda-subs.jar.

It also adds a buildscript method that makes a submodule be compiled against
this (reduced) JDK. All submodules that are intended to be used on Android
should call this method:
  useReducedJdk()
Such submodules won't be able to use languge APIs (packages, classes and methods)
that are not available on Android.