Estimated gas provider
In ContractGasProvider: move deprecation to methods that take
'contractFunc' parameter (it is not relevant to gas use-case);
add 'getGasLimit' that takes all known relevant data about the
transaction (nonce is not visible on-chain and cannot affect gas limit);
add IOException to methods' signatures.
In Contract: remove deprecated 'setGasPrice' and 'getGasPrice' as they
cannot be used without breaking the functionality of a Gas Provider.