Add support for priv_distributeRawTransaction API method (https://besu.hyperledger.org/en/stable/Reference/API-Methods/#priv_distributerawtransaction)
PrivacyTransactionManager.java

Added new method signAndDistribute which takes a raw private transaction signs it and the distibutes to private transaction rceipis using the BEsu API method privDistributeRawTransaction

Added new method sign to support signing but not distributing a raw provate transaction.

Refactored the existing sendTransaction methods to reuse the sign method via signAndSend, changes are transparent to existing users.

The sign method resolves a bug where if a chainId of ChainIDLong.NONE is used the eeaSendRawTransaction fails with a confusing error.