Minor Fix to adapt with different chains.
Long ago, I was working using this library to work with matic chain (Polygon). Apparently, It seems that in the JSON, they send a long value in string format. So it kept throwing error in the code. I use this patched version in my code, but i thought i should make this change on original code and propose a pull request so that others don't have to go through the same issue.